refactor(apps::communication): refactored requests & responses
Added operation codes and error codes to fully integrate with the arduino communication infrastructure. Modified the byte order of addresses and opcode. Defined specific masks for address in order to specify if either a LoRa packet or an Iridium Packet will be sent. Still need to add parameter to constructor when the IridiumCommunicator is implemented